+++ internal/frontend/bridge-gui/bridge-gui/qml/Proton/TextArea.qml
@@ -1,4 +1,4 @@
-// Copyright (c) 2024 Proton AG
+// Copyright (c) 2025 Proton AG
 // This file is part of Proton Mail Bridge.
 // Proton Mail Bridge is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -366,5 +366,6 @@ FocusScope {
     Proton.ContextMenu {
         parentObject: root
         colorScheme: root.colorScheme
+        readOnly: control.readOnly
     }
 }