+++ utils/bridge_app_version.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-# Copyright (c) 2024 Proton AG
+# Copyright (c) 2025 Proton AG
 #
 # This file is part of Proton Mail Bridge.
 #