+++ utils/vault-editor/main.go
@@ -1,6 +1,6 @@
 //go:build debug
 
-// Copyright (c) 2024 Proton AG
+// Copyright (c) 2025 Proton AG
 //
 // This file is part of Proton Mail Bridge.
 //