+++ Changelog.md
@@ -24,6 +24,12 @@ Changelog [format](http://keepachangelog.com/en/1.0.0/)
 * BRIDGE-258: Fixed issue with draft updates and sending during synchronization.
 
 
+## Erasmus Bridge 3.15.1
+
+### Changed
+* BRIDGE-281: Disable keychain test on macOS.
+
+
 ## Erasmus Bridge 3.15.0
 
 ### Added